Automate rm of repo-*86* when commiting "any" pkg
This commit integrates part of the functionality of commitpkg2any
in the standard commitpkg script. If the package is for arch=any
and there exists ../repos/$repo-{i686,x86_64}, then these are
removed.

NB: I believe commitpkg2any should be removed, as it encourages
converting a pkg to arch=any with the same $pkgver. It is better
practice to wait for a new version before releasing the pkg
with arch=any. It also saves bandwidth!

Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>