arch-nspawn: include all host mirrors in nspawn
https://lists.archlinux.org/pipermail/arch-projects/2018-July/004970.html

The original scenario was a host with a local caching service. As
arch-nspawn would only choose the very first mirror in the host pacman
configuration, only this caching service would be available in the
nspawn container. Avoid this by including all mirrors on the host to
the container.

To avoid unnecessary mangling with sed, pacconf --raw is used instead
of pacman-conf. This introduces a dependency of pacutils on devtools,
but could be changed if required.

Signed-off-by: Alad Wenter <alad@archlinux.org>