Rework the way mount args are built. This fixes #29 and probably #14.
Move an improved version of the "while read -r ..." loop inside the
build_mount_args function and remove the first for loop in function.

In the while loop, the repo name and architecture in a server url are
no longer replaced by sed with the $repo and $arch placeholder strings.
When a local repo url is added to the cache_dirs array and later, when
it is used to build --bind-ro args it represents a valid filesystem path.
systemd-nspwan won't throw a "Failed to stat [...]" error anymore.
An extra if was removed and more friendly variable named were used.

The first for loop was removed since the improved while loop now handles
all the cases.