arch-nspawn: obsolete nosetarch option
setarch is mainly used for building i686 packages on x86_64, while it
disturbs building ARM packages on x86_64, so the nosetarch option (-s)
was introduced. But there is a better way that we could always precheck
the setarch support list, and do it only when available.

This could fix the partial implement of setarch refered on #41 .