makechrootpkg: fix an infinite loop when installing package with -I
When installing with the '-I' option a package which in turn has
a dependency with multiple providers, makechrootpkg enters an
infinite loop. This happens because we are using the 'pacman -U'
command piped with 'yes y'. The 'yes y' command outputs only 'y'
lines, but when installing a dependency package with multiple
providers pacman expects a number as the input. Being such,
makechrootpkg outputs this endlessly:

    Enter a number (default=1): error: invalid number: y

Using the pacman '--noconfirm' option without piping 'yes y' fixes
the issue.