fix(version): do not prompt for input on terminal during check
In certain scenarios, like when an URL is checked that asks for
credentials or other kind of input, the process blocks and waits for
stdin input. This isn't particularly useful during scripted execution
and in our cases also just an indicator for like a none existing GitHub
repository.

To avoid this situation, disable git terminal prompts.

Component: pkgctl version check
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>