fix(drop-from-repo): avoid error when updating remote head
We can only update the local ref to the remote head from main to master
once we have actually pushed the master branch to the remote. If we do
this too early, the call will simply error out.

Instead, only update the local head for the configured remote after we
have successfully renamed the branch and pushed it to the AUR.

Fixes #212

Component: pkgctl aur drop-from-repo
Signed-off-by: Levente Polyak <anthraxx@archlinux.org>